Introduce API server with DuckDB integration and improved ETL logic:
- Added an HTTP API server with endpoints for querying candle data (`/v1/data`) and executing SQL (`/v1/sql`) using Gorilla/Mux.
- Integrated DuckDB for efficient querying of Parquet files, including dynamic table creation.
- Improved ETL error handling with proper channel closure for CSV and Parquet streams.
- Enhanced CSV to struct mapping with a dedicated test for reflection-based parsing.
- Updated `main.go` to serve the API and removed legacy ETL execution logic.

+++ pkg/binance/model.go
@@ -253,7 +253,7 @@ type Kline struct {
 	NumberOfTrades int64   `csv:"8"`
 	TakerBuyVolume float64 `csv:"9"`
 	TakerBuyQuote  float64 `csv:"10"`
-	Ignore         int64   `csv:"11"`
+	Ignore         float64 `csv:"11"`
 }
 
 type ParquetKline struct {
